forager_server/requirements.txt
django==3.1.1
pillow
google-cloud-storage
